+++ server/src/feed/service/feed.service.ts
@@ -15,9 +15,7 @@ import {
   FeedTrendResponseDto,
 } from '../dto/feed-response.dto';
 import { RedisService } from '../../common/redis/redis.service';
-import { Cron, CronExpression } from '@nestjs/schedule';
 import { EventEmitter2 } from '@nestjs/event-emitter';
-import * as _ from 'lodash';
 import {
   SearchFeedReq,
   SearchFeedRes,
@@ -88,32 +86,6 @@ export class FeedService {
     return FeedTrendResponseDto.toFeedTrendResponseDtoArray(
       trendFeeds.filter((feed) => feed !== null),
     );
-  }
-
-  @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
-  async resetTrendTable() {
-    await this.redisService.redisClient.del(redisKeys.FEED_TREND_KEY);
-  }
-
-  @Cron(CronExpression.EVERY_30_SECONDS)
-  async analyzeTrend() {
-    const [originTrend, nowTrend] = await Promise.all([
-      this.redisService.redisClient.lrange(
-        redisKeys.FEED_ORIGIN_TREND_KEY,
-        0,
-        3,
-      ),
-      this.redisService.redisClient.zrevrange(redisKeys.FEED_TREND_KEY, 0, 3),
-    ]);
-
-    if (!_.isEqual(originTrend, nowTrend)) {
-      const redisPipeline = this.redisService.redisClient.pipeline();
-      redisPipeline.del(redisKeys.FEED_ORIGIN_TREND_KEY);
-      redisPipeline.rpush(redisKeys.FEED_ORIGIN_TREND_KEY, ...nowTrend);
-      await redisPipeline.exec();
-      const trendFeeds = await this.readTrendFeedList();
-      this.eventService.emit('ranking-update', trendFeeds);
-    }
   }
 
   async searchFeedList(searchFeedReq: SearchFeedReq) {
@@ -200,16 +172,6 @@ export class FeedService {
     tomorrow.setDate(tomorrow.getDate() + 1);
     tomorrow.setHours(0, 0, 0, 0);
     return tomorrow;
-  }
-
-  @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
-  async resetIpTable() {
-    const redis = this.redisService.redisClient;
-    const keys = await redis.keys(redisKeys.FEED_ALL_IP_KEY);
-
-    if (keys.length > 0) {
-      await redis.del(...keys);
-    }
   }
 
   async readRecentFeedList() {
